# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Soccer Connect API",
    "description": "Backend API for Soccer Connect - GTA soccer community platform",
    "version": "1.0.0",
    "contact": {
      "name": "Soccer Connect Team"
    }
  },
  "servers": [
    {
      "url": "https://soccer-connect-api.vercel.app/api",
      "description": "Production"
    },
    {
      "url": "http://localhost:8046/api",
      "description": "Local Development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
...
  "paths": {
...
    "/fields": {
      "get": {
        "tags": ["Fields"],
        "summary": "Get all fields with filters and pagination",
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 20 } },
          { "name": "surface", "in": "query", "schema": { "type": "string" } },
          { "name": "type", "in": "query", "schema": { "type": "string" } },
          { "name": "search", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "List of fields" }
        }
      },
      "post": {
        "tags": ["Fields"],
        "summary": "Create a new field (admin only)",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "201": { "description": "Field created" }
        }
      }
    },
    "/fields/nearby": {
      "get": {
        "tags": ["Fields"],
        "summary": "Get fields near a location",
        "parameters": [
          { "name": "lat", "in": "query", "required": true, "schema": { "type": "number" } },
          { "name": "lng", "in": "query", "required": true, "schema": { "type": "number" } },
          { "name": "radius", "in": "query", "schema": { "type": "number", "default": 10 } }
        ],
        "responses": {
          "200": { "description": "Nearby fields" }
        }
      }
    },
    "/fields/{id}": {
      "get": {
        "tags": ["Fields"],
        "summary": "Get single field",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Field details" }
        }
      },
      "put": {
        "tags": ["Fields"],
        "summary": "Update field (admin only)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Field updated" }
        }
      },
      "delete": {
        "tags": ["Fields"],
        "summary": "Delete field (admin only)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Field deleted" }
        }
      }
    },
    "/fields/{id}/availability": {
      "get": {
        "tags": ["Fields"],
        "summary": "Get field availability for a specific date",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "date", "in": "query", "required": true, "schema": { "type": "string", "format": "date" } }
        ],
        "responses": {
          "200": { "description": "Availability slots" }
        }
      }
    },
...
  }
}